commit_id for local installation from source refere to id at installtime
Ofc the assumption that the commit at a given time was the one used for the installation is error-prone since it could be on a different branch. But it is better than using the current commit for an installation that bundled files in the past (there is no installing from the future).